backend/requirements.txt
# Core API
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-multipart==0.0.7

# Database
sqlalchemy==2.0.25
sqlmodel==0.0.14
asyncpg==0.29.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-dotenv==1.2.1
PyJWT==2.10.1

# Scrapers & Playwright
playwright==1.41.1
playwright-stealth==1.0.6
beautifulsoup4==4.12.3
requests==2.31.0
dateparser==1.2.2
urllib3==2.5.0

# Tasks & Scheduling
APScheduler==3.10.4

# Payments
razorpay==2.0.0

# PDF & QR Generation
reportlab==4.0.7
qrcode==7.4.2

# Email
mailjet-rest==1.3.4
email-validator>=2.0.0
fastapi-mail==1.4.1

# Utilities
jinja2==3.1.6
aiohttp==3.13.3
certifi==2025.11.12

# AI & GenAI
openai==1.12.0
langchain>=0.1.0
langchain-core==0.1.23
langchain-google-genai==0.0.9
langchain-groq==0.0.1
duckduckgo-search>=6.0.0
Pillow==10.2.0